tests/requirements/dj42_cms41.txt
-r ./requirements_base.txt

Django>=4.2,<5.0
django-cms>=4.1,<4.2

djangocms-versioning>=2.0.1
djangocms-alias>=2.0.0
